@@ -13,22 +13,21 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "connect-mongo": "^3.2.0",
+    "connect-mongo": "^5.0.0",
     "cors": "^2.8.5",
-    "crypto-js": "^4.0.0",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "express-session": "^1.17.1",
+    "crypto-js": "^4.1.1",
+    "dotenv": "^16.0.3",
+    "express": "^4.18.2",
+    "express-session": "^1.17.3",
     "json2csv": "^5.0.6",
-    "mongoose": "^5.12.1",
-    "morgan": "^1.10.0",
-    "node-fetch": "^2.6.0",
-    "nodemon": "^2.0.4",
-    "passport": "^0.4.1",
-    "passport-discord": "^0.1.4",
-    "pm2": "^4.5.5"
+    "mongoose": "^7.2.0",
+    "morgan": "^1.10.0",    
+    "axios": "^1.4.0",
+    "nodemon": "^2.0.22",
+    "passport": "^0.6.0",
+    "passport-discord": "^0.1.4"
   },
   "devDependencies": {
-    "eslint": "^7.7.0"
+    "eslint": "^8.41.0"
   }
 }
