@@ -2,13 +2,13 @@ require('dotenv').config();
 require('./strategies/discord');
 
 const express = require('express');
-const path = require('path');
+// const path = require('path');
 const passport = require('passport');
 const morgan = require('morgan');
 const mongoose = require('mongoose');
 const session = require('express-session');
 const cors = require('cors')
-const Store = require('connect-mongo')(session);
+const MongoStore = require('connect-mongo');
 const PORT = process.env.PORT || 3000;
 const app = express();
 const routes = require('./routes');
@@ -17,8 +17,6 @@ const { getGuildCount } = require('./utils/api');
 mongoose.connect(process.env.MONGO_LOGIN, {
 	useNewUrlParser: true,
 	useUnifiedTopology: true,
-	useFindAndModify: false,
-	useCreateIndex: true
 });
 
 app.use(session({
@@ -26,10 +24,10 @@ app.use(session({
 	cookie: { maxAge: 60000 * 60 * 24 },
 	resave: false,
 	saveUninitialized: false,
-	store: new Store({ mongooseConnection: mongoose.connection })
+	store: MongoStore.create({ mongoUrl: process.env.MONGO_LOGIN })
 }));
 app.use(cors({
-	origin: ['http://localhost:3000'],
+	origin: ['http://localhost:3000', 'http://localhost:3000/'],
 	credentials: true
 }));
 app.use(passport.initialize());
@@ -37,10 +35,11 @@ app.use(passport.session());
 app.use(express.json());
 app.use(morgan('combined'));
 
-app.get('/logout', function(request, response){
-	console.log('logout');
-	request.logout();
-	response.redirect(process.env.BASE_URL);
+app.get('/logout', function(request, response, next){
+	request.logout(function(err) {
+		if (err) return next(err);
+		response.redirect(process.env.REACT_APP_BASE_URL);
+	});
 });
 
 app.use('/api', routes);
@@ -51,16 +50,20 @@ app.get('/count', async function(req, resp) {
 	resp.send({count});
 });
 
-//Required for deployment to production
-app.use(express.static(path.join(__dirname)));
+// app.use(express.static(path.join(__dirname, '..', process.env.WEB_FOLDER)));
 
-app.get('/*', function (req, res) {
-	console.log(path.join(__dirname, '..', process.env.WEB_FOLDER, 'index.html'));
-	console.log(path.join(__dirname, '..', 'index.html'));
-	res.sendFile(path.join(__dirname, 'index.html'));
-});
+// app.get('/*', function (req, res) {
+// 	res.sendFile(path.join(__dirname, '..', process.env.WEB_FOLDER, 'index.html'));
+// });
+//Required for deployment to production
+// app.use(express.static(path.join(__dirname)));
+
+// app.get('/*', function (req, res) {
+// 	console.log(path.join(__dirname, '..', process.env.WEB_FOLDER, 'index.html'));
+// 	console.log(path.join(__dirname, '..', 'index.html'));
+// 	res.sendFile(path.join(__dirname, 'index.html'));
+// });
 
 app.listen(PORT, () => {
-	console.log(process.env);
 	console.log(`Running on port ${PORT}`)
 });
