@@ -6,8 +6,8 @@ require('dotenv').config();
 router.get('/discord', passport.authenticate('discord'));
 
 //Failure redirect will send back to our homepage if user clicks cancel.
-router.get('/discord/redirect', passport.authenticate('discord', { failureRedirect: process.env.BASE_URL || '/' }), (request, response) => {
-	response.redirect(`${process.env.BASE_URL}menu`)
+router.get('/discord/redirect', passport.authenticate('discord', { failureRedirect: process.env.REACT_APP_BASE_URL || '/' }), (request, response) => {
+	response.redirect(`${process.env.REACT_APP_BASE_URL}menu`)
 });
 
 router.get('/', (request, response) => {
