@@ -20,7 +20,7 @@ passport.use(
 	new DiscordStrategy({
 		clientID: process.env.DASHBOARD_CLIENT_ID,
 		clientSecret: process.env.DASHBOARD_CLIENT_SECRET,
-		callbackURL: process.env.BASE_URL + process.env.DASHBOARD_CALLBACK_URL.substring(1),
+		callbackURL: process.env.APP_BASE_URL + process.env.DASHBOARD_CALLBACK_URL.substring(1),
 		scope: ['identify', 'guilds']
 	}, async (accessToken, refreshToken, profile, done) => {
 		const { id, username, discriminator, avatar, guilds } = profile;
